Appointment Controller,Service,Repository

+++ src/main/java/com/realestate/repository/AppointmentRepository.java
@@ -3,5 +3,6 @@ package com.realestate.repository;
 import com.realestate.model.Appointment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+public interface AppointmentRepository
+        extends JpaRepository<Appointment, Integer> {
 }
